+++ src/app/pager/pager-standalone.demo.ts
@@ -17,14 +17,16 @@ export class PagerStandaloneDemoComponent {
     disablePreviousButton: false,
     disableNextButton: false,
 
-    firstPageTooltip   : 'click to got to the first page of records',
-    lastPageTooltip    : 'click to got to the last page of records',
+    firstPageTooltip: 'click to got to the first page of records',
+    lastPageTooltip: 'click to got to the last page of records',
     previousPageTooltip: 'click to got to the previous page of records',
-    nextPageTooltip    : 'click to got to the last page of records',
+    nextPageTooltip: 'click to got to the last page of records',
+
+    hidePageSelectorInput: false,
 
     hidePageSizeSelector: false,
     useSmallPageSizeSelector: false,
-    pageSizeMenuSettings: {attachToBody: false},
+    pageSizeMenuSettings: { attachToBody: false },
 
     pageSize: 10,
     pageSizes: [5, 10, 15, 20],